# docs/api.yml
openapi: 3.1.0
info:
  title: Server Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: "Refresh Access Token"
      security: [ ]
      responses:
        200:
          description: Access token refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          headers:
            Set-Cookie:
              description: Http only cookies contains refresh token
              schema:
                type: string
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Response:
      type: object
      properties:
        message:
          type: string
...
    AuthenticationResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
...
security:
  - BearerAuth: []